Baselines v6.0.0-rc.1

+++ SapHana.Odbc.LinqService/Tests/xUpdate/TruncateTableTests/Tests.xUpdate.TruncateTableTests.TruncateTableTest(SapHana.Odbc.LinqService).sql
@@ -8,8 +8,8 @@ BeforeExecute
 
 CREATE COLUMN TABLE "TestTrun"
 (
-	"ID"     Integer NOT NULL,
-	"Field1" Decimal NOT NULL,
+	"ID"     Integer         NOT NULL,
+	"Field1" Decimal(38, 10) NOT NULL,
 
 	PRIMARY KEY ("ID")
 )
