+++ SqlServer.2012.MS/Tests/xUpdate/TruncateTableTests/Tests.xUpdate.TruncateTableTests.TruncateTableTest(SqlServer.2012.MS).sql
@@ -9,8 +9,8 @@ BeforeExecute
 
 CREATE TABLE [TestTrun]
 (
-	[ID]     Int     NOT NULL,
-	[Field1] Decimal NOT NULL,
+	[ID]     Int             NOT NULL,
+	[Field1] Decimal(18, 10) NOT NULL,
 
 	CONSTRAINT [PK_TestTrun] PRIMARY KEY CLUSTERED ([ID])
 )
