+++ Sybase.Managed.LinqService/Tests/Linq/TypesTests/Tests.Linq.TypesTests.Issue4469Test2(Sybase.Managed.LinqService,True).sql
@@ -2,7 +2,7 @@
 -- Sybase.Managed Sybase
 
 SELECT TOP 2
-	CAST([v].[Integer] AS DECIMAL) / 33,
+	CAST([v].[Integer] AS DECIMAL(18, 10)) / 33,
 	[v].[Decimal] / 33,
 	[v].[Double] / 33
 FROM
